Everything works (Apparently)

+++ RowDaBoat-x64barebones-d4e1c147f975/Userland/SampleCodeModule/shell.h
@@ -3,6 +3,6 @@ void startGraphicMode(const char* input, int length);
 void printTime();
 void graphicCuadraticFunction();
 void plotQuadraticFunction(int a, int b, int c);
-void screenToWorldCoordinates(float *x, float *y, int unitWidth, int unitHeight);
-void worldToScreenCoordinates(float *x, float *y, int unitWidth, int unitHeight);
+void screenToWorldCoordinates(double *x, double *y, int unitWidth, int unitHeight);
+void worldToScreenCoordinates(double *x, double *y, int unitWidth, int unitHeight);
 void colorCrossAxis(int unitWidth, int unitHeight);
